backend/requirements.txt
flask==2.0.2
flask-cors==3.0.10
waitress==2.0.0
kubernetes==21.7.0
pyyaml==6.0
lxml==4.7.1
pyinstaller
